# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Coworking Management System",
      "description": "API documentation for the Coworking Management System",
      "version": "1.0.0"
    },
...
    "paths": {
...
      "/contracts": {
        "get": {
          "tags": ["contracts"],
          "summary": "Retrieve all contracts",
          "responses": {
            "200": {
              "description": "List of all contracts",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": { "$ref": "#/components/schemas/Contract" }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": ["contracts"],
          "summary": "Create a new contract",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreateContractDto" }
              }
            }
          },
          "responses": {
            "201": {
              "description": "The created contract",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/Contract" }
                }
              }
            }
          }
        }
      },
      "/contracts/{id}": {
        "get": {
          "tags": ["contracts"],
          "summary": "Retrieve a contract by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": { "type": "string" }
            }
          ],
          "responses": {
            "200": {
              "description": "The contract",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/Contract" }
                }
              }
            },
            "404": {
              "description": "Contract not found"
            }
          }
        },
        "put": {
          "tags": ["contracts"],
          "summary": "Update a contract by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": { "type": "string" }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UpdateContractDto" }
              }
            }
          },
          "responses": {
            "200": {
              "description": "The updated contract",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/Contract" }
                }
              }
            },
            "404": {
              "description": "Contract not found"
            }
          }
        },
        "delete": {
          "tags": ["contracts"],
          "summary": "Delete a contract by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": { "type": "string" }
            }
          ],
          "responses": {
            "200": {
              "description": "The deleted contract",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/Contract" }
                }
              }
            },
            "404": {
              "description": "Contract not found"
            }
          }
        }
      },
      "/contract-values": {
        "get": {
          "tags": ["contract-values"],
          "summary": "Retrieve all contract values",
          "responses": {
            "200": {
              "description": "List of all contract values",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": { "$ref": "#/components/schemas/ContractValue" }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": ["contract-values"],
          "summary": "Create a new contract value",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreateContractValueDto" }
              }
            }
          },
          "responses": {
            "201": {
              "description": "The created contract value",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ContractValue" }
                }
              }
            }
          }
        }
      },
      "/contract-values/{id}": {
        "get": {
          "tags": ["contract-values"],
          "summary": "Retrieve a contract value by ID",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": { "type": "string" }
            }
          ],
          "responses": {
            "200": {
              "description": "The contract value",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ContractValue" }
                }
              }
            },
            "404": {
              "description": "Contract value not found"
            }
          }
        }
      },
      "/payments": {
        "get": {
          "tags": ["payments"],
          "summary": "Retrieve all payments",
          "responses": {
            "200": {
              "description": "List of all payments",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": { "$ref": "#/components/schemas/Payment" }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": ["payments"],
          "summary": "Register a payment",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreatePaymentDto" }
              }
            }
          },
          "responses": {
            "201": {
              "description": "The registered payment",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/Payment" }
                }
              }
            }
          }
        }
      },
      "/payments/create": {
      "post": {
        "tags": ["payments"],
        "summary": "Create a new payment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreatePaymentDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created payment",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Payment" }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    },
    "components": {
      "schemas": {
...
        "Contract": {
          "type": "object",
          "properties": {
            "id": { "type": "string" },
            "coworkerId": { "type": "string" },
            "nuc": { "type": "string" },
            "startDate": { "type": "string", "format": "date-time" },
            "endDate": { "type": "string", "format": "date-time" },
            "contractType": { "type": "string" },
            "status": { "type": "string" },
            "previousContractId": { "type": "string" },
            "roomId": { "type": "string" }
          }
        },
        "CreateContractDto": {
          "type": "object",
          "properties": {
            "coworkerId": { "type": "string" },
            "nuc": { "type": "string" },
            "startDate": { "type": "string", "format": "date-time" },
            "endDate": { "type": "string", "format": "date-time" },
            "contractType": { "type": "string" },
            "status": { "type": "string" },
            "amount": { "type": "number" },
            "previousContractId": { "type": "string" },
            "roomId": { "type": "string" }
          },
          "required": ["coworkerId", "nuc", "startDate", "endDate", "contractType", "status", "amount"]
        },
        "UpdateContractDto": {
          "type": "object",
          "properties": {
            "coworkerId": { "type": "string" },
            "nuc": { "type": "string" },
            "startDate": { "type": "string", "format": "date-time" },
            "endDate": { "type": "string", "format": "date-time" },
            "contractType": { "type": "string" },
            "status": { "type": "string" },
            "amount": { "type": "number" },
            "previousContractId": { "type": "string" },
            "roomId": { "type": "string" }
          }
        },
        "ContractValue": {
          "type": "object",
          "properties": {
            "id": { "type": "string" },
            "contractId": { "type": "string" },
            "createdAt": { "type": "string", "format": "date-time" },
            "amount": { "type": "number" }
          }
        },
        "CreateContractValueDto": {
          "type": "object",
          "properties": {
            "contractId": { "type": "string" },
            "amount": { "type": "number" }
          },
          "required": ["contractId", "amount"]
        },
        "Payment": {
          "type": "object",
          "properties": {
            "id": { "type": "string" },
            "invoiceId": { "type": "string" },
            "paymentDate": { "type": "string", "format": "date-time" },
            "amountPaid": { "type": "number" }
          }
        },
        "CreatePaymentDto": {
          "type": "object",
          "properties": {
            "invoiceId": { "type": "string" },
            "paymentDate": { "type": "string", "format": "date-time" },
            "amountPaid": { "type": "number" }
          },
          "required": ["invoiceId", "paymentDate", "amountPaid"]
        },
...
      }
    }
  }